Exttract generic piop tests (temporary solution)
